% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{CleanText}
\alias{CleanText}
\title{A utility function to clean text}
\usage{
CleanText(text)
}
\arguments{
\item{text}{A piece of text to be cleaned. Not a vector.}
}
\value{
Cleaned text.
}
\description{
A utility function to clean text
}

