% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{CountDifficultWords}
\alias{CountDifficultWords}
\title{Count occurrences of "difficult" words}
\usage{
CountDifficultWords(text, easy_word_list)
}
\arguments{
\item{text}{A piece (segment) of text to be counted.}

\item{easy_word_list}{A list of easy words.}
}
\value{
The count of difficult words
}
\description{
Given a list of "easy" words, count occurrences of "difficult" words---
words not in the provided easy word list. This is a utility function for
calculating P_Lex.
}

