% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{GetSegments}
\alias{GetSegments}
\title{Segment a piece of text}
\usage{
GetSegments(text, len = 10)
}
\arguments{
\item{text}{The piece of text to be segmented.}

\item{len}{The length of each segment. In other words, the number of words in each segment.}
}
\value{
Text segments, represented by a character vector.
}
\description{
Split a piece of text into segments of n words. This is a utility function for calculating P_Lex.
}

