% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/GslAwl.R
\docType{data}
\name{GslAwl}
\alias{GslAwl}
\title{The GSL/AWL word lists}
\format{A list of three vectors, each corresponding to one of the following word lists:
first 1,000 words, second 1,000 words, and Academic Word list.}
\usage{
GslAwl
}
\description{
This dataset containing the General Service List (GSL) developed by by Michael West (1953)
and the Academic Word List developed by Coxhead (2000).
}
\keyword{datasets}

