% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{LexicalFrequency}
\alias{LexicalFrequency}
\title{Compute lexical frequency}
\usage{
LexicalFrequency(text, wordlist = "GslAwl")
}
\arguments{
\item{text}{Text to be analyzed.}

\item{wordlist}{Word list to be used. Only GSL/AWL is currently supported.}
}
\value{
An integer vector, containing frequencies for different types of
words defined by the specified word list.
}
\description{
Compute lexical frequency for a piece of text.
}
\examples{
LexicalFrequency(some_awesome_text)
}

