% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{LoadGslAwl}
\alias{LoadGslAwl}
\title{Load GSL/AWL}
\usage{
LoadGslAwl()
}
\value{
A list of word lists. Each list is represented as a character vector.
}
\description{
Load the General Service List (GSL) developed by by Michael West (1953)
and the Academic Word List developed by Coxhead (2000).
}
\examples{
lists = LoadGslAwl()
}

