% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/LexFreq_function.R
\name{P_Lex}
\alias{P_Lex}
\title{Calculate P_Lex for a piece of text}
\usage{
P_Lex(text, len = 10, wordlist = "GslAwl")
}
\arguments{
\item{text}{A piece of text to be analyzed.}

\item{len}{The number of words contained in each segment.}

\item{wordlist}{The word list the computation is based. Default: `GslAwl`.
Only GSL/AWL is currently supported. More will be added in the future.}
}
\value{
A lamda score, with a higher score denoting more difficult text.
}
\description{
Calculate P_Lex for a piece of text. Technically, P_Lex is the lambda score
of the distribution of counts of difficult words in segments of a piece of text.
}
\references{
Meara, P., & Bell, H. (2001). P_Lex: A Simple and Effective Way of Describing the
lexical Characteristics of Short L2 Tests. Prospect, 16(3), 5–19.
}

